refactor(domain): PR-2 schema.py 976 → studio/domain/ 拆分
0.11.0 studio/ 重构 8 PR 系列第二刀。原 studio/schema.py 单文件 976 行
（含 5 个 BaseModel + 2 个 migration + 1 个 helper）拆到 studio/domain/
子包；studio/schema.py 改为兼容垫片。**零业务改动**，仅结构搬运。

新文件：
- studio/domain/__init__.py    re-export 全部公共名
- studio/domain/common.py      _meta() helper + AttentionBackend + GROUP_ORDER
- studio/domain/migrations.py  migrate_legacy_attention + migrate_legacy_save_keys
- studio/domain/training.py    TrainingConfig（643 行单类，本次保持单类单文件；
                                 拆 sub-mixin 属于行为改动，留独立 PR）
- studio/domain/lora.py        LoraEntry
- studio/domain/xy_matrix.py   XYAxisType + XYAxisSpec + XYMatrixSpec + _check_axis_values
- studio/domain/generate.py    GenerateConfig
- studio/domain/reg.py         RegAiConfig

studio/schema.py: 976 → 38 行 re-export 垫片；所有 `from studio.schema import X`
仍工作（含 _meta / _check_axis_values 等私有名，避免破坏 runtime/ 和 tests/）。

回归：
- PR-1 安全网（route snapshot / import smoke / invariants）全绿
- 全测试套 1586 passed（baseline 1578 + 新增 8 个 domain/* import smoke 用例）
- 0 failed

注：本次 TrainingConfig 保持 643 行单文件单类。Agent A 原方案的 5 个 mixin
拆分（OptimizerMixin / NetworkMixin / NoiseMixin / ...）涉及 pydantic 多重继承
+ 字段顺序 / model_dump 输出变化，属行为改动，需独立 PR + 严格回归测试。
